# asset_openapi.yaml
openapi: "3.2.0"
info:
  title: AssetPilot API
  version: "4.110.1"
  description: |
    個人資產管理平台 AssetPilot 的完整 API Schema。
    供 Cloudflare API Shield Schema Validation 使用。
    認證方式：JWT 儲存於 httpOnly Cookie（token），除 /api/auth/* 與 /api/config 外均需認證。
    記帳 API 可用 X-Ledger-Id header 或 ledgerId query 選擇帳本，未指定則使用個人帳本。
    每次請求皆驗證登入者的帳本成員資格；viewer 僅可 GET/HEAD/OPTIONS。
    個人偏好、投資、登入憑證及 MCP/API Token/LINE 整合不採用共享帳本選擇。
servers:
  - url: https://asset.shao.one
    description: Production

# ─────────────────────────────────────────
# 標籤定義（對應 Cloudflare API Shield Endpoint Labels）
# ─────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  schemas:
...
    AccountInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        initialBalance:
          type: number
        icon:
          type: string
        currency:
          type: string
        accountType:
          type: string
        excludeFromTotal:
          type: boolean
        linkedBankId:
          type: string

    # ── 交易 ──
...
security:
  - cookieAuth: []

# ─────────────────────────────────────────
# 路由定義
# ─────────────────────────────────────────
paths:
...
  /api/account/delete:
    post:
      operationId: deleteAccount
      tags: [Account]
      summary: 永久刪除帳號及所有資料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmEmail]
              properties:
                confirmEmail:
                  type: string
                  format: email
      responses:
        "200":
          description: 帳號已刪除
...
  /api/accounts:
    get:
      operationId: getAccounts
      tags: [Accounts]
      summary: 取得帳戶清單（含餘額）
      responses:
        "200":
          description: 帳戶陣列
    post:
      operationId: createAccount
      tags: [Accounts]
      summary: 新增帳戶
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "200":
          description: 新增成功
  /api/accounts/{id}:
    put:
      operationId: updateAccount
      tags: [Accounts]
      summary: 更新帳戶
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                initialBalance:
                  type: number
                icon:
                  type: string
                currency:
                  type: string
                accountType:
                  type: string
                excludeFromTotal:
                  type: boolean
                linkedBankId:
                  type: string
      responses:
        "200":
          description: 更新成功
    delete:
      operationId: deleteAccountEntry
      tags: [Accounts]
      summary: 刪除帳戶
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 刪除成功
...